Separate code for init
