+++ piperider_cli/initializer.py
@@ -1,0 +1,98 @@
+import os
+
+from rich.console import Console
+
+from piperider_cli import clone_directory
+from piperider_cli.datasource import DataSource
+from piperider_cli.configuration import Configuration, \
+    PIPERIDER_WORKSPACE_NAME, \
+    PIPERIDER_CONFIG_PATH, \
+    PIPERIDER_CREDENTIALS_PATH
+
+
+def _is_piperider_workspace_exist(workspace_path: str) -> bool:
+    if not os.path.exists(workspace_path):
+        return False
+    elif not os.path.exists(os.path.join(workspace_path, 'config.yml')):
+        return False
+
+    return True
+
+
+def _generate_piperider_workspace() -> bool:
+    from piperider_cli import data
+    init_template_dir = os.path.join(os.path.dirname(data.__file__), 'piperider-init-template')
+    working_dir = os.path.join(os.getcwd(), PIPERIDER_WORKSPACE_NAME)
+
+    if _is_piperider_workspace_exist(working_dir) is False:
+        clone_directory(init_template_dir, working_dir)
+        # prepare .gitignore file
+        os.rename(os.path.join(working_dir, 'gitignore'), os.path.join(working_dir, '.gitignore'))
+        return True
+    else:
+        # Skip if workspace already exists
+        return False
+
+
+def _ask_user_update_credentials(ds: DataSource):
+    console = Console()
+    console.print(f'\nPlease enter the following fields for {ds.type_name}')
+    return ds.ask_credential()
+
+
+def _ask_user_input_datasource(config: Configuration = None):
+    console = Console()
+    if config is None:
+        cls, name = DataSource.ask()
+        ds: DataSource = cls(name=name)
+        config = Configuration([ds])
+        if _ask_user_update_credentials(ds):
+            config.dump(PIPERIDER_CONFIG_PATH)
+            config.dump_credentials(PIPERIDER_CREDENTIALS_PATH)
+    else:
+        if len(config.dataSources) == 1:
+            ds = config.dataSources[0]
+        else:
+            ds = config.ask_for_datasource()
+        if not ds.credential:
+            console.print(
+                f'[[bold yellow]Warning[/bold yellow]] No credential found for \'{ds.type_name}\' datasource \'{ds.name}\'')
+            if _ask_user_update_credentials(ds):
+                config.dump_credentials(PIPERIDER_CREDENTIALS_PATH)
+
+    ds.show_installation_information()
+    return config
+
+
+def _inherit_datasource_from_dbt_project(dbt_project_path, dbt_profiles_dir=None,
+                                         config: Configuration = None) -> bool:
+    config = Configuration.from_dbt_project(dbt_project_path, dbt_profiles_dir)
+    config.dump(PIPERIDER_CONFIG_PATH)
+
+    return config
+
+
+def _generate_configuration(dbt_project_path=None, dbt_profiles_dir=None):
+    """
+    :param dbt_project_path:
+    :return: Configuration object
+    """
+    try:
+        config = Configuration.load()
+    except Exception:
+        config = None
+    if dbt_project_path is None:
+        return _ask_user_input_datasource(config=config)
+
+    return _inherit_datasource_from_dbt_project(dbt_project_path, dbt_profiles_dir)
+
+
+class Initializer():
+    def exec(dbt_project_path=None, dbt_profiles_dir=None):
+        console = Console()
+        if _generate_piperider_workspace() is False:
+            console.print('[bold green]Piperider workspace already exist[/bold green] ')
+
+        # get Configuration object from dbt or user created configuration
+        configuration = _generate_configuration(dbt_project_path, dbt_profiles_dir)
+        return configuration
