+++ piperider_cli/workspace.py
@@ -14,106 +14,15 @@ from piperider_cli import clone_directory, convert_to_tzlocal, datetime_to_str, 
 from piperider_cli.assertion_engine import AssertionEngine, ValidationResult
 from piperider_cli.assertion_engine.recommender import RECOMMENDED_ASSERTION_TAG
 from piperider_cli.compare_report import CompareReport
-from piperider_cli.configuration import Configuration, PIPERIDER_WORKSPACE_NAME, PIPERIDER_CONFIG_PATH, \
-    PIPERIDER_CREDENTIALS_PATH
 from piperider_cli.datasource import DataSource
 from piperider_cli.error import PipeRiderCredentialError, PipeRiderNoProfilingResultError
 from piperider_cli.profiler import Profiler
-
-PIPERIDER_OUTPUT_PATH = os.path.join(os.getcwd(), PIPERIDER_WORKSPACE_NAME, 'outputs')
-PIPERIDER_REPORT_PATH = os.path.join(os.getcwd(), PIPERIDER_WORKSPACE_NAME, 'reports')
-PIPERIDER_COMPARISON_PATH = os.path.join(os.getcwd(), PIPERIDER_WORKSPACE_NAME, 'comparisons')
-
-CONSOLE_MSG_PASS = '[bold green]✅ PASS[/bold green]\n'
-CONSOLE_MSG_FAIL = '[bold red]😱 FAILED[/bold red]\n'
-CONSOLE_MSG_ALL_SET = '[bold]🎉 You are all set![/bold]\n'
-
-
-def _is_piperider_workspace_exist(workspace_path: str) -> bool:
-    if not os.path.exists(workspace_path):
-        return False
-    elif not os.path.exists(os.path.join(workspace_path, 'config.yml')):
-        return False
-
-    return True
-
-
-def _generate_piperider_workspace() -> bool:
-    from piperider_cli import data
-    init_template_dir = os.path.join(os.path.dirname(data.__file__), 'piperider-init-template')
-    working_dir = os.path.join(os.getcwd(), PIPERIDER_WORKSPACE_NAME)
-
-    if _is_piperider_workspace_exist(working_dir) is False:
-        clone_directory(init_template_dir, working_dir)
-        # prepare .gitignore file
-        os.rename(os.path.join(working_dir, 'gitignore'), os.path.join(working_dir, '.gitignore'))
-        return True
-    else:
-        # Skip if workspace already exists
-        return False
-
-
-def _ask_user_update_credentials(ds: DataSource):
-    console = Console()
-    console.print(f'\nPlease enter the following fields for {ds.type_name}')
-    return ds.ask_credential()
-
-
-def _ask_user_input_datasource(config: Configuration = None):
-    console = Console()
-    if config is None:
-        cls, name = DataSource.ask()
-        ds: DataSource = cls(name=name)
-        config = Configuration([ds])
-        if _ask_user_update_credentials(ds):
-            config.dump(PIPERIDER_CONFIG_PATH)
-            config.dump_credentials(PIPERIDER_CREDENTIALS_PATH)
-    else:
-        if len(config.dataSources) == 1:
-            ds = config.dataSources[0]
-        else:
-            ds = config.ask_for_datasource()
-        if not ds.credential:
-            console.print(
-                f'[[bold yellow]Warning[/bold yellow]] No credential found for \'{ds.type_name}\' datasource \'{ds.name}\'')
-            if _ask_user_update_credentials(ds):
-                config.dump_credentials(PIPERIDER_CREDENTIALS_PATH)
-
-    ds.show_installation_information()
-    return config
-
-
-def _inherit_datasource_from_dbt_project(dbt_project_path, dbt_profiles_dir=None,
-                                         config: Configuration = None) -> bool:
-    config = Configuration.from_dbt_project(dbt_project_path, dbt_profiles_dir)
-    config.dump(PIPERIDER_CONFIG_PATH)
-
-    return config
-
-
-def _generate_configuration(dbt_project_path=None, dbt_profiles_dir=None):
-    """
-    :param dbt_project_path:
-    :return: Configuration object
-    """
-    try:
-        config = Configuration.load()
-    except Exception:
-        config = None
-    if dbt_project_path is None:
-        return _ask_user_input_datasource(config=config)
-
-    return _inherit_datasource_from_dbt_project(dbt_project_path, dbt_profiles_dir)
-
-
-def init(dbt_project_path=None, dbt_profiles_dir=None):
-    console = Console()
-    if _generate_piperider_workspace() is False:
-        console.print('[bold green]Piperider workspace already exist[/bold green] ')
-
-    # get Configuration object from dbt or user created configuration
-    configuration = _generate_configuration(dbt_project_path, dbt_profiles_dir)
-    return configuration
+from piperider_cli.configuration import Configuration, \
+    PIPERIDER_WORKSPACE_NAME, \
+    PIPERIDER_CONFIG_PATH, \
+    PIPERIDER_CREDENTIALS_PATH, \
+    PIPERIDER_OUTPUT_PATH, \
+    PIPERIDER_COMPARISON_PATH
 
 
 def _agreed_to_run_recommended_assertions(console: Console, interactive: bool):
@@ -600,7 +509,6 @@ def generate_recommended_assertions(input=None, interaction=True):
     console.rule('Generated Recommended Assertions')
     for f in recommended_assertions:
         console.print(f'[bold green]Recommended Assertion[/bold green]: {f}')
-    pass
 
 
 def _append_descriptions(profile_result):
@@ -655,5 +563,3 @@ def compare_report(a=None, b=None):
     # TODO for debugging intermediate data, remove this
     with open(os.path.join(dir, 'comparison_data.json'), 'w') as f:
         f.write(comparison_data.to_json())
-
-    pass
